Add optional day/night schedule, remove redundant 307 hive call, resolve undefined variable.
Instructions for enabling and configuring the day/night cycle option are at the top of the file. It is very easy for server owners to configure. Instead of performing a 307 hive call, this file uses the date stored in ServerCurrentTime that is updated every 60 seconds in sched_event. dayz_storeTimeDate is used in server_sendToClient to respond to client date requests sent by player_monitor. It was undefined on the server. The saving grace is the redundancy in publicVariable sends by the server. The client was bound to get the correct date eventually.