Add and update backapacks
-	5 new backpacks
-	19 new textures for the Czech backpack
-	2 new textures for the coyote backpack
-	reconfigured weapon count of all old backpacks
-	30 new backpack types included an upgraded version

This adds 5 new backpacks. 4 from the ice apo mod 1 from Arma 2 ACR.
This also adds many different skins for the Czech backpack thx @lennard91 and 2 new skins for the coyote backpack from the KSK Mod.
I’ve reconfigured the weapon count of the old backpacks. A weapon can be a pistol with 5 slots too. So, a Large Gunbag can hold 6 weapons or 12 pistols now. This is more intuitive as it was before. 6x10 = 60, 12x5=60 The backpack has still the same gear count as before. I removed the display info in the trader menu how many weapons a backpack can hold. This information is obsolete with this commit because a backpack with a gear count of 20 can hold 4 weapons now and it could be confusing for players if they consider only large weapons but no pistols.

We added 15 reconfigured backpack types with DZE1 as indicator. Those new backpack types have an upgraded version with DZE2 as indicator. The gear count, the names and descriptions of all new backpacks got changed.

This commit was made by @seelenapparat, @DeVloek, @airwavesman