Change used ammo for weapons
This commitment changes the type of ammo that can be used in silenced and non-silenced weapons. SD and non-SD ammo can currently be used in all weapon variations. We know this is possible and normal in real life, but arma weapons and ammo work in an unrealistic way. In the current state a SD weapon is almost no longer necessary since all normal weapons can be used with SD ammo to avoid triggering zombies or AIs. This is the same way that regular ammo in SD weapons works. This triggers AI's and zombies but not players. With this commitment we restore the actual and intended purpose of silenced weapons. SD ammo and SD weapons will have more relevance after this change. All silenced weapons can now only be used with SD ammo, and all non-silenced weapons can only be used with regular (non-SD) ammo.