Add fix for AI targeting exploit.
This fix prevents players from exploiting the change clothes function to reset AI targeting. Players must either kill all of the near AI or move 800 meters away to switch clothes. Credit to Arrakis for pointing out this exploit. Credit to @morgoth0 for suggesting the use of the getHideFrom command. Reorganized this file a bit to place the checks up top and removed unused code.