Change scope and type of configs from name to numbers
This got discussed earlier. Since epoch got merged with DayZ Mod we have a different name for the scope and type of configs.  A proper config should use numbers for those entries. This is the same how arma configs are written.  The basicDefines.hpp has all the needed information stored.

Also, this removes the weight entries from the magazines since we have them in a separate folder, so admins can edit them.