Update player_checkStealth.sqf
Several functions check if the player is inside a building including the new weather functions. The DayZ Mod devs introduced the dayz_inside variable in fn_temperatur but never finished the code to utilize it. Since player_checkStealth is called continuously in the scheduler we can set the variable here and use it in other functions.