Add coins on zombies
Coins on zombies and the value can be handled with ZSC_ZombieCoins now. The coins for zombies will be handled directly in check wallet.

Also add a PVE check to disable check wallet if the player is not the owner of the body or in the group with the owner of the body.