Fix exorbitant damage from chainsaws and sledge hammers
Reduces the damage from chainsaws and sledge hammers to prevent the very fast destruction of cinder and metal buildings.

This is the only good way to prevent the high damage on cinder and metal. Eventhandlers for damage are not a good idea and can affect the server performace if to many are set.