Revert object_pickupAction back to execVM
I was not aware of the acutal behavior of the init eventHandler. The init eventhandler sends all the given arguments to the player before the mpframework got even load on the player, so it is basicly the same like the publicvariables. This means that the function object_pickupAction will be always nil because it gets defined far later after the init evenhandler fired already.