Streamline Vehicle Eventhandlers
Almost all eventhandlers getting set already when creating the base class of the vehicles like AllVehicles. So it is not needed to set it after that again. Until a few exceptions I removed or added the needed vehicle eventhandlers.

All helicopters have a fired eventhandler now as it should be. So finally if a helicopters shoots all crew members are in combat. This is the same if flares getting fired.

The fired evenhandler will be only executed for players now no longer for the server.

Note: The CSJ Gyro addon getting loaded very early so those addons need all eventhandlers set.