Update weapon accuarcy
Some weapons are inaccurate at a certain distance. Mostly because of wrong config entries from some weapon addons ive overtook. All weapons should be accurate now. This was not only a problem of the sd snipers.
This also updates the scope optics of the Mk17 Sniper SD and the VSS.
Thx to Jimmy, Goose and Th3-Hunter333 for letting me know.