Update snappoints.hpp
Adjust z position to center of metal and glass floors. Adjust the top of the cinder walls so they line up with the top of the cinder gate when stacked 2 high. Add additional snap points to the 4x metal floors.

The 3d models of the wood floors need adjustment so the x and y axis are at the center of the models.