Add Arma 2 Static Weapon as Epoch Classes
This adds for every type of arma 2 static weapon an epoch version for it. This also adds to every epoch static weapon a magazine class so it can be build easily. The new weapons can found under: DayZ Epoch Static Weapons in the editor. Server admins have to add Battleye Filters for most of the static weapons. Since adding it to all server would weaken also servers which do not use those new static weapons.