Slightly increased SD ammo speed
SD weapons with Stanag and AK ammo and an PSO or ACOG scope are inaccurate at a smaller distance. This happends due to the low speed of the sd ammo itself. Only a slightly higher bullet speed solves this. If we set the zeroing of the scopes lower we break the distance indicators of those scopes.